+++ DeeployTest/Platforms/Generic/main.c
@@ -50,22 +50,37 @@ int main() {
 
   int32_t tot_err = 0;
   uint32_t tot = 0;
-  float32_t diff;
-  float32_t expected, actual;
+  OUTPUTTYPE diff;
+  OUTPUTTYPE expected, actual;
+
   for (uint32_t buf = 0; buf < DeeployNetwork_num_outputs; buf++) {
-    tot += DeeployNetwork_outputs_bytes[buf] / sizeof(float32_t);
-    for (uint32_t i = 0;
-         i < DeeployNetwork_outputs_bytes[buf] / sizeof(float32_t); i++) {
-      expected = ((float32_t *)testOutputVector[buf])[i];
-      actual = ((float32_t *)DeeployNetwork_outputs[buf])[i];
+    tot += DeeployNetwork_outputs_bytes[buf] / sizeof(OUTPUTTYPE);
+    for (uint32_t i = 0; i < DeeployNetwork_outputs_bytes[buf] / sizeof(OUTPUTTYPE); i++) {
+      expected = ((OUTPUTTYPE *)testOutputVector[buf])[i];
+      actual = ((OUTPUTTYPE *)DeeployNetwork_outputs[buf])[i];
       diff = expected - actual;
 
-      if ((diff < 0 ? -diff : diff) > 1e-5) {
+#if OUTPUTTYPE == float32_t
+      // RUNWANG: Allow margin of error for float32_t
+      if ((diff < -1e-4) || (diff > 1e-4))
+      {
         tot_err += 1;
-        printf("Expected: %10.6f  ", expected);
-        printf("Actual: %10.6f  ", actual);
-        printf("Diff: %10.6f at Index %12u in Output %u\r\n", diff, i, buf);
+        printf("Expected: %10.6f  ", (float)expected);
+        printf("Actual: %10.6f  ", (float)actual);
+        printf("Diff: %10.6f at Index %12u in Output %u\r\n", (float)diff, i, buf);
       }
+#elif OUTPUTTYPE == int32_t
+      // RUNWANG: No margin for integer comparison
+      if (diff != 0)
+      {
+        tot_err += 1;
+        printf("Expected: %4d  ", expected);
+        printf("Actual: %4d  ", actual);
+        printf("Diff: %4d at Index %12u in Output %u\r\n", diff, i, buf);
+      }
+#else
+#error "Unsupported OUTPUTTYPE"
+#endif
     }
   }